@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.4.31'
+    ext.kotlin_version = '1.5.21'
     ext.dokka_version = '1.5.0'
 
     repositories {
@@ -9,7 +9,7 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.2'
+        classpath 'com.android.tools.build:gradle:7.0.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath "org.jetbrains.dokka:dokka-gradle-plugin:$dokka_version"
         classpath "org.jetbrains.dokka:kotlin-as-java-plugin:$dokka_version"
@@ -21,8 +21,7 @@ buildscript {
 
 allprojects {
     repositories {
-        maven { url = "https://dl.bintray.com/kotlin/dokka" }
-        mavenCentral()
         google()
+        mavenCentral()
     }
 }
