@@ -4,8 +4,8 @@ import android.content.res.AssetManager;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.net.Uri;
-import android.support.test.filters.MediumTest;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.filters.MediumTest;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import com.squareup.moshi.Moshi;
 
@@ -51,13 +51,14 @@ import java.util.Map;
 
 import bolts.Task;
 
-import static android.support.test.InstrumentationRegistry.getTargetContext;
+import static androidx.test.core.app.ApplicationProvider.getApplicationContext;
 import static net.gini.android.Utils.CHARSET_UTF8;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
+import static org.mockito.ArgumentMatchers.nullable;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.never;
@@ -78,7 +79,7 @@ public class DocumentTaskManagerTest {
     @Before
     public void setUp() {
         // https://code.google.com/p/dexmaker/issues/detail?id=2
-        System.setProperty("dexmaker.dexcache", getTargetContext().getCacheDir().getPath());
+        System.setProperty("dexmaker.dexcache", getApplicationContext().getCacheDir().getPath());
 
         mApiCommunicator = Mockito.mock(ApiCommunicator.class);
         mSessionManager = Mockito.mock(SessionManager.class);
@@ -91,7 +92,7 @@ public class DocumentTaskManagerTest {
     }
 
     private Bitmap createBitmap() throws IOException {
-        AssetManager assetManager = getTargetContext().getResources().getAssets();
+        AssetManager assetManager = getApplicationContext().getResources().getAssets();
 
         InputStream inputStream;
         inputStream = assetManager.open("yoda.jpg");
@@ -103,14 +104,14 @@ public class DocumentTaskManagerTest {
     }
 
     private byte[] createByteArray(String filename) throws IOException {
-        AssetManager assetManager = getTargetContext().getResources().getAssets();
+        AssetManager assetManager = getApplicationContext().getResources().getAssets();
 
         InputStream inputStream = assetManager.open(filename);
         return TestUtils.createByteArray(inputStream);
     }
 
     private JSONObject readJSONFile(final String filename) throws IOException, JSONException {
-        InputStream inputStream = getTargetContext().getResources().getAssets().open(filename);
+        InputStream inputStream = getApplicationContext().getResources().getAssets().open(filename);
         int size = inputStream.available();
         byte[] buffer = new byte[size];
         @SuppressWarnings("unused")
@@ -120,7 +121,7 @@ public class DocumentTaskManagerTest {
     }
 
     private JSONArray readJSONArrayFile(final String filename) throws IOException, JSONException {
-        InputStream inputStream = getTargetContext().getResources().getAssets().open(filename);
+        InputStream inputStream = getApplicationContext().getResources().getAssets().open(filename);
         int size = inputStream.available();
         byte[] buffer = new byte[size];
         @SuppressWarnings("unused")
@@ -134,7 +135,7 @@ public class DocumentTaskManagerTest {
         BufferedReader inputStreamReader = null;
         try {
             final AssetManager assetManager =
-                    getTargetContext().getResources().getAssets();
+                    getApplicationContext().getResources().getAssets();
             inputStreamReader = new BufferedReader(
                     new InputStreamReader(assetManager.open("document-template.json"), CHARSET_UTF8));
             StringBuilder stringBuilder = new StringBuilder();
@@ -246,7 +247,7 @@ public class DocumentTaskManagerTest {
     public void testThatCreateDocumentResolvesToDocument() throws IOException, JSONException, InterruptedException {
         final Uri createdDocumentUri = Uri.parse("https://pay-api.gini.net/documents/1234");
         when(mApiCommunicator.uploadDocument(any(byte[].class), any(String.class), any(String.class), any(String.class),
-                any(Session.class), any(DocumentMetadata.class)))
+                any(Session.class), nullable(DocumentMetadata.class)))
                 .thenReturn(Task.forResult(createdDocumentUri));
         when(mApiCommunicator.getDocument(eq(createdDocumentUri), any(Session.class))).thenReturn(
                 createDocumentJSONTask("1234"));
@@ -262,7 +263,7 @@ public class DocumentTaskManagerTest {
             throws IOException, JSONException, InterruptedException {
         final Uri createdDocumentUri = Uri.parse("https://pay-api.gini.net/documents/1234");
         when(mApiCommunicator.uploadDocument(any(byte[].class), any(String.class), any(String.class), any(String.class),
-                any(Session.class), any(DocumentMetadata.class)))
+                any(Session.class), nullable(DocumentMetadata.class)))
                 .thenReturn(Task.forResult(Uri.parse("https://pay-api.gini.net/documents/1234")));
         when(mApiCommunicator.getDocument(eq(createdDocumentUri), any(Session.class))).thenReturn(
                 createDocumentJSONTask("1234"));
@@ -272,7 +273,7 @@ public class DocumentTaskManagerTest {
 
         verify(mApiCommunicator)
                 .uploadDocument(any(byte[].class), eq("application/vnd.gini.v1.partial+jpeg"), eq("foobar.jpg"), eq("Invoice"),
-                        eq(mSession), any(DocumentMetadata.class));
+                        eq(mSession), nullable(DocumentMetadata.class));
     }
 
     @Test
@@ -280,7 +281,7 @@ public class DocumentTaskManagerTest {
         final Uri createdDocumentUri = Uri.parse("https://pay-api.gini.net/documents/1234");
         when(mApiCommunicator.uploadDocument(any(byte[].class), any(String.class),
                 any(String.class), any(String.class),
-                any(Session.class), any(DocumentMetadata.class)))
+                any(Session.class), nullable(DocumentMetadata.class)))
                 .thenReturn(Task.forResult(Uri.parse("https://pay-api.gini.net/documents/1234")));
         when(mApiCommunicator.getDocument(eq(createdDocumentUri), any(Session.class))).thenReturn(
                 createDocumentJSONTask("1234"));
@@ -293,7 +294,7 @@ public class DocumentTaskManagerTest {
                 .uploadDocument(eq(document),
                         eq("application/vnd.gini.v1.partial+jpeg"), eq("foobar.jpg"),
                         eq("Invoice"),
-                        eq(mSession), any(DocumentMetadata.class));
+                        eq(mSession), nullable(DocumentMetadata.class));
     }
 
     @Test
@@ -301,7 +302,7 @@ public class DocumentTaskManagerTest {
         final Uri createdDocumentUri = Uri.parse("https://pay-api.gini.net/documents/1234");
         when(mApiCommunicator.uploadDocument(any(byte[].class), any(String.class),
                 any(String.class), any(String.class),
-                any(Session.class), any(DocumentMetadata.class)))
+                any(Session.class), nullable(DocumentMetadata.class)))
                 .thenReturn(Task.forResult(Uri.parse("https://pay-api.gini.net/documents/1234")));
         when(mApiCommunicator.getDocument(eq(createdDocumentUri), any(Session.class))).thenReturn(
                 createDocumentJSONTask("1234"));
@@ -316,7 +317,7 @@ public class DocumentTaskManagerTest {
                 .uploadDocument(any(byte[].class),
                         eq("application/vnd.gini.v1.composite+json"), eq((String) null),
                         eq("Invoice"),
-                        eq(mSession), any(DocumentMetadata.class));
+                        eq(mSession), nullable(DocumentMetadata.class));
     }
 
     @Test
@@ -324,7 +325,7 @@ public class DocumentTaskManagerTest {
         final Uri createdDocumentUri = Uri.parse("https://pay-api.gini.net/documents/1234");
         when(mApiCommunicator.uploadDocument(any(byte[].class), any(String.class),
                 any(String.class), any(String.class),
-                any(Session.class), any(DocumentMetadata.class)))
+                any(Session.class), nullable(DocumentMetadata.class)))
                 .thenReturn(Task.forResult(Uri.parse("https://pay-api.gini.net/documents/1234")));
         when(mApiCommunicator.getDocument(eq(createdDocumentUri), any(Session.class))).thenReturn(
                 createDocumentJSONTask("1234"));
@@ -346,7 +347,7 @@ public class DocumentTaskManagerTest {
                 .uploadDocument(eq(jsonBytes),
                         eq("application/vnd.gini.v1.composite+json"), eq((String) null),
                         eq("Invoice"),
-                        eq(mSession), any(DocumentMetadata.class));
+                        eq(mSession), nullable(DocumentMetadata.class));
     }
 
     @Test
@@ -354,7 +355,7 @@ public class DocumentTaskManagerTest {
         final Uri createdDocumentUri = Uri.parse("https://pay-api.gini.net/documents/1234");
         when(mApiCommunicator.uploadDocument(any(byte[].class), any(String.class),
                 any(String.class), any(String.class),
-                any(Session.class), any(DocumentMetadata.class)))
+                any(Session.class), nullable(DocumentMetadata.class)))
                 .thenReturn(Task.forResult(Uri.parse("https://pay-api.gini.net/documents/1234")));
         when(mApiCommunicator.getDocument(eq(createdDocumentUri), any(Session.class))).thenReturn(
                 createDocumentJSONTask("1234"));
@@ -376,7 +377,7 @@ public class DocumentTaskManagerTest {
                 .uploadDocument(eq(jsonBytes),
                         eq("application/vnd.gini.v1.composite+json"), eq((String) null),
                         eq("Invoice"),
-                        eq(mSession), any(DocumentMetadata.class));
+                        eq(mSession), nullable(DocumentMetadata.class));
     }
 
     @Test
@@ -384,7 +385,7 @@ public class DocumentTaskManagerTest {
         final Uri createdDocumentUri = Uri.parse("https://pay-api.gini.net/documents/1234");
         when(mApiCommunicator.uploadDocument(any(byte[].class), any(String.class),
                 any(String.class), any(String.class),
-                any(Session.class), any(DocumentMetadata.class)))
+                any(Session.class), nullable(DocumentMetadata.class)))
                 .thenReturn(Task.forResult(Uri.parse("https://pay-api.gini.net/documents/1234")));
         when(mApiCommunicator.getDocument(eq(createdDocumentUri), any(Session.class))).thenReturn(
                 createDocumentJSONTask("1234"));
@@ -406,7 +407,7 @@ public class DocumentTaskManagerTest {
                 .uploadDocument(eq(jsonBytes),
                         eq("application/vnd.gini.v1.composite+json"), eq((String) null),
                         eq("Invoice"),
-                        eq(mSession), any(DocumentMetadata.class));
+                        eq(mSession), nullable(DocumentMetadata.class));
     }
 
     @Test
